pyproject.toml
[project]
name = "hybrid-search-practice"
version = "0.1.0"
description = "Hybrid Search Practice - A playground for experimenting with multiple RAG/Hybrid Search approaches (PageIndex, LightRAG, HiRAG, etc.)"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "openai>=1.12.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "aiofiles>=23.2.0",
    # Hybrid Search modules dependencies
    "pageindex>=0.2.5",
    "lightrag-hku>=1.4.10",          # LightRAG - Graph-based dual-level retrieval
    # PageIndex submodule dependencies
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.11.0",
    "pyyaml>=6.0.0",
    "pillow>=10.0.0",
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    # Graph and ML dependencies (for LightRAG/HiRAG)
    "networkx>=3.0",
    "numpy>=1.24",
    # Semantic BM25 + Vector Search (Hybrid)
    "qdrant-client[fastembed]>=1.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    # Code quality tools
    "ruff>=0.11.0", # Python linter and formatter (replaces black, flake8, isort)
    "mypy>=1.15.0", # Static type checker
    "pre-commit>=4.0.0", # Git hooks framework
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# ==================== Ruff Configuration ====================
# Ruff: 极速 Python linter 和 formatter
# 文档: https://docs.astral.sh/ruff/
[tool.ruff]
target-version = "py313"  # Python 3.13
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "ASYNC", # flake8-async
]
ignore = [
    "D100",  # Missing docstring in public module (模块级文档字符串非强制)
    "D104",  # Missing docstring in public package
    "ASYNC230",  # Async functions should not open files (项目中允许阻塞 I/O)
    "B008",  # Do not perform function call in argument defaults (FastAPI 常见模式)
    "E501",  # Line too long (多行字符串不易缩短)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

# 针对测试文件的宽松规则
[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D103",  # Missing docstring in public function (测试函数不要求文档字符串)
    "D102",  # Missing docstring in public method (测试方法不要求文档字符串)
    "ASYNC230",  # Async functions should not open files (测试中允许)
    "ASYNC240",  # Async functions should not use pathlib (测试中允许)
]
"verify_lightrag.py" = [
    "F401",  # Unused imports (verify 脚本中的导入用于检查)
]
"lib/hirag/*" = [
    "ALL",  # 忽略 HiRAG submodule 的所有检查
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ==================== MyPy Configuration ====================
# MyPy: 静态类型检查器
# 文档: https://mypy.readthedocs.io/
[tool.mypy]
python_version = "3.13"
warn_return_any = false  # 可选导入模式会产生大量此警告
warn_unused_configs = true
disallow_untyped_defs = false  # 项目暂时允许无类型注解
disallow_incomplete_defs = false
ignore_missing_imports = true  # 忽略第三方库的类型缺失
show_error_codes = true
show_column_numbers = true
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false

# 忽略特定模块的类型检查（第三方库无类型支持）
[[tool.mypy.overrides]]
module = [
    "pageindex.*",
    "lightrag.*",
    "hirag",
    "hirag.*",
    "fitz",
]
ignore_missing_imports = true

# 忽略包含复杂类型问题的文件
[[tool.mypy.overrides]]
module = [
    "config",
    "main",
    "models.database",
    "models.document_store",
    "services.document_service",
    "services.global_search_service",
    "services.legacy.*",
    "tests.*",
    "verify_lightrag",
    "lib.hirag_wrapper",
]
ignore_errors = true  # 这些文件包含 SQLAlchemy 和可选导入的复杂类型问题
